Explicit cast to truncate integers
Make the implict cast to a smaller integer size explicit so that reviewer can
see it.

Note that this cast safely doesn't alter the value since it is in a branch where
that value has been tested to be smaller than UINT32_MAX.

You can test with

   nix-build -A c --builders '' --arg nixpkgs '(import <nixpkgs> {}).pkgsi686Linux' --argstr env clangStdenv

+++ jets.c
@@ -1117,7 +1117,7 @@ static void div_mod_96_64(uint_fast32_t *q, uint_fast64_t *r,
   uint_fast64_t estQ = ah / bh;
 
   /* Precondition 1 guarentees Q is 32-bits, if estQ is greater than UINT32_MAX, then reduce our initial estimated quotient to UINT32_MAX. */
-  *q = estQ <= UINT32_MAX ? estQ : UINT32_MAX;
+  *q = estQ <= UINT32_MAX ? (uint_fast32_t)estQ : UINT32_MAX;
 
   /* *q * bh <= estQ * bh <= ah */
   uint_fast64_t rh = ah - 1u * *q * bh;
